refine decorator
